Modified CreateTodo for Class
