+++ controllers/todos/GetTodoController.ts
@@ -1,0 +1,14 @@
+// import { TodoRepository } from "../../repositories/TodoRepository";
+// import { TodoController } from "../../controllers/todos/CreateTodoController";
+
+// const todoRepository = new TodoRepository();
+// const todoController = new TodoController(todoRepository);
+
+// import { TodoRepository } from "../../repositories/TodoRepository";
+
+// list(req: Request, res: Response) {
+//     const todos = this.repository.list()
+
+//     return res.status(200).json(todos)
+//   }
+// export class GetTodosController {}
