+++ package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "start": "nodemon app.ts",
     "test": "jest"
   },
   "author": "",
