--- routers/todos.ts
@@ -1,7 +1,0 @@
-import express from "express";
-const router = express.Router();
-import { postTodo } from "../controllers/todos";
-
-router.route("/").post(postTodo);
-
-export default router;
