+++ routers/todos/list.ts
@@ -1,0 +1,5 @@
+// import { router } from ".";
+// import
+// router.route("/").get((req, res) => todoController.list(req, res));
+
+// export default router;
